+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - Tooltip Module UX: Indicate that certain frames are required to be active for functionality.
 - Tooltip Module UX: Seperated tooltip options into Time Frame / System Stats Frame.
 - Added ChangeLog.
+
 # V5.2: 16 / 07 / 2024
-- Fix: Instance Difficulty not hiding when you close config with Instance Difficulty Tab active.
-- Feature: Raid Lockouts will now be abbreviated on the tooltip.
+- Fixed Instance Difficulty Frame not hiding on config window close.
+- Abbreviate raid titles in tooltip.
+- Tidied up `FetchVaultOptions()`, and added comments so I remember what it does.