+++ CHANGELOG.md
@@ -13,3 +13,4 @@
 - Renamed `AbbrRaidTitles` -> `AbbrInstances` for maintainability.
 - Added `WoWProject` tags to Friends List.
 - `RequestRaidInfo` added to `ENCOUNTER_END` check for better functionality with lockouts and tooltips.
+- Better abbreviations for raid difficulties for tooltips.