+++ CHANGELOG.md
@@ -15,3 +15,4 @@
 - `RequestRaidInfo` added to `ENCOUNTER_END` check for better functionality with lockouts and tooltips.
 - Better abbreviations for raid difficulties for tooltips.
 - Added Textures for `Online`, `AFK` and `DND` Friend Tooltips.
+- Fixed Tooltip Information Spacing.